Pull in changes from main repo

+++ modules/javafx.graphics/src/main/java/com/sun/glass/ui/monocle/LinuxEventBuffer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2020, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -90,7 +90,7 @@ class LinuxEventBuffer {
      */
     synchronized boolean put(ByteBuffer event) throws
             InterruptedException {
-        boolean isSync = event.getInt(eventStruct.getTypeIndex()) == 0
+        boolean isSync = event.getShort(eventStruct.getTypeIndex()) == 0
                 && event.getInt(eventStruct.getValueIndex()) == 0;
         while (bb.limit() - bb.position() < event.limit()) {
             // Block if bb is full. This should be the
